+++ projects/packages/my-jetpack/tests/php/test-wpcom-products.php
@@ -211,10 +211,11 @@ class Test_Wpcom_Products extends TestCase {
 		remove_filter( 'pre_http_request', array( $this, 'mock_success_response' ) );
 
 		$expected = array(
-			'currency_code'   => 'BRL',
-			'full_price'      => 4.9,
-			'discount_price'  => 2.45,
-			'coupon_discount' => 50,
+			'currency_code'         => 'BRL',
+			'full_price'            => 4.9,
+			'discount_price'        => 2.45,
+			'is_introductory_offer' => false,
+			'coupon_discount'       => 50,
 		);
 
 		$this->assertSame( $expected, $product_price );