+++ projects/packages/videopress/src/client/admin/hooks/use-analytics-tracks/index.ts
@@ -11,12 +11,11 @@ import { useAnalyticsTracksProps } from './types';
 
 const useAnalyticsTracks = ( {
 	pageViewEventName,
-	pageViewNamespace = 'jetpack_videopress',
-	pageViewSuffix = 'page_view',
 	pageViewEventProperties = {},
 }: useAnalyticsTracksProps ) => {
 	const { isUserConnected, isRegistered, userConnectionData } = useConnection();
-	const { login, ID } = userConnectionData.currentUser?.wpcomUser || {};
+	const { blogId } = userConnectionData?.currentUser || {};
+	const { login, ID } = userConnectionData?.currentUser?.wpcomUser || {};
 
 	// Tracks
 	const { tracks } = jetpackAnalytics;
@@ -38,9 +37,14 @@ const useAnalyticsTracks = ( {
 			callback = typeof properties === 'function' ? properties : callback;
 			properties = typeof properties === 'function' ? {} : properties;
 
+			// Populate event props with blog ID when it's available.
+			if ( blogId ) {
+				properties.blog_id = blogId;
+			}
+
 			return () => recordEventAsync( eventName, properties ).then( callback );
 		},
-		[ recordEventAsync ]
+		[ recordEventAsync, blogId ]
 	);
 
 	// Initialize Analytics identifying the user.
@@ -57,22 +61,22 @@ const useAnalyticsTracks = ( {
 	 * It's considered a page view event
 	 * when the component is mounted.
 	 */
-	const pageViewEvent = pageViewEventName
-		? `${ pageViewNamespace }_${ pageViewEventName }_${ pageViewSuffix }`
-		: null;
-
 	useEffect( () => {
 		// Also, only run if the site is registered.
 		if ( ! isRegistered ) {
 			return;
 		}
 
-		if ( ! pageViewEvent ) {
+		if ( ! pageViewEventName ) {
 			return;
 		}
 
-		recordEvent( pageViewEvent, pageViewEventProperties );
-	}, [] );
+		if ( blogId ) {
+			pageViewEventProperties.blog_id = blogId;
+		}
+
+		recordEvent( pageViewEventName, pageViewEventProperties );
+	}, [ blogId ] );
 
 	return {
 		recordEvent: recordEventAsync,