+++ projects/plugins/boost/app/assets/src/js/stores/config.ts
@@ -1,8 +1,6 @@
 import { writable } from 'svelte/store';
 import api from '../api/api';
-import { saveGetStarted } from '../api/get-started';
 
-// eslint-disable-next-line camelcase
 const { subscribe, update } = writable( Jetpack_Boost );
 
 async function refresh(): Promise< void > {
@@ -13,7 +11,6 @@ async function refresh(): Promise< void > {
 	} );
 }
 
-// eslint-disable-next-line camelcase
 const dismissedPopOutStore = writable( Jetpack_Boost.dismissedScorePrompts );
 
 export const dismissedPopOuts = {
@@ -31,7 +28,6 @@ export const markGetStartedComplete = () => {
 			getStarted: false,
 		},
 	} ) );
-	return saveGetStarted();
 };
 
 export default {