+++ projects/plugins/jetpack/extensions/blocks/contact-form/variations.js
@@ -1,9 +1,10 @@
 import { isSimpleSite } from '@automattic/jetpack-shared-extension-utils';
 import { Path } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-import { compact } from 'lodash';
+import { compact, map } from 'lodash';
 import { getIconColor } from '../../shared/block-icons';
 import renderMaterialIcon from '../../shared/render-material-icon';
+import { childBlocks } from './child-blocks';
 
 const defaultBlockStyling = {
 	style: {
@@ -33,9 +34,9 @@ const variations = compact( [
 			'-4 -4 32 32'
 		),
 		innerBlocks: [
-			[ 'jetpack/field-name', { required: true } ],
-			[ 'jetpack/field-email', { required: true } ],
-			[ 'jetpack/field-textarea', {} ],
+			[ 'jetpack/field-name', { required: true, label: __( 'Name', 'jetpack' ) } ],
+			[ 'jetpack/field-email', { required: true, label: __( 'Email', 'jetpack' ) } ],
+			[ 'jetpack/field-textarea', { label: __( 'Message', 'jetpack' ) } ],
 			[
 				'jetpack/button',
 				{
@@ -66,8 +67,8 @@ const variations = compact( [
 			'-6 -6 92 92'
 		),
 		innerBlocks: [
-			[ 'jetpack/field-name', { required: true } ],
-			[ 'jetpack/field-email', { required: true } ],
+			[ 'jetpack/field-name', { required: true, label: __( 'Name', 'jetpack' ) } ],
+			[ 'jetpack/field-email', { required: true, label: __( 'Email', 'jetpack' ) } ],
 			[ 'jetpack/field-consent', {} ],
 			[
 				'jetpack/button',
@@ -96,8 +97,8 @@ const variations = compact( [
 			'-4 -3 32 32'
 		),
 		innerBlocks: [
-			[ 'jetpack/field-name', { required: true } ],
-			[ 'jetpack/field-email', { required: true } ],
+			[ 'jetpack/field-name', { required: true, label: __( 'Name', 'jetpack' ) } ],
+			[ 'jetpack/field-email', { required: true, label: __( 'Email', 'jetpack' ) } ],
 			[
 				'jetpack/field-radio',
 				{
@@ -135,9 +136,9 @@ const variations = compact( [
 			'-4 -3 32 32'
 		),
 		innerBlocks: [
-			[ 'jetpack/field-name', { required: true } ],
-			[ 'jetpack/field-email', { required: true } ],
-			[ 'jetpack/field-telephone', { label: __( 'Phone Number', 'jetpack' ) } ],
+			[ 'jetpack/field-name', { required: true, label: __( 'Name', 'jetpack' ) } ],
+			[ 'jetpack/field-email', { required: true, label: __( 'Email', 'jetpack' ) } ],
+			[ 'jetpack/field-telephone', { label: __( 'Phone', 'jetpack' ) } ],
 			[
 				'jetpack/field-select',
 				{
@@ -180,9 +181,9 @@ const variations = compact( [
 			'-4 -3 32 32'
 		),
 		innerBlocks: [
-			[ 'jetpack/field-name', { required: true } ],
-			[ 'jetpack/field-email', { required: true } ],
-			[ 'jetpack/field-telephone', { required: true } ],
+			[ 'jetpack/field-name', { required: true, label: __( 'Name', 'jetpack' ) } ],
+			[ 'jetpack/field-email', { required: true, label: __( 'Email', 'jetpack' ) } ],
+			[ 'jetpack/field-telephone', { required: true, label: __( 'Phone', 'jetpack' ) } ],
 			[ 'jetpack/field-date', { label: __( 'Date', 'jetpack' ), required: true } ],
 			[
 				'jetpack/field-radio',
@@ -221,8 +222,8 @@ const variations = compact( [
 			'-4 -3 32 32'
 		),
 		innerBlocks: [
-			[ 'jetpack/field-name', { required: true } ],
-			[ 'jetpack/field-email', { required: true } ],
+			[ 'jetpack/field-name', { required: true, label: __( 'Name', 'jetpack' ) } ],
+			[ 'jetpack/field-email', { required: true, label: __( 'Email', 'jetpack' ) } ],
 			[
 				'jetpack/field-radio',
 				{
@@ -252,6 +253,30 @@ const variations = compact( [
 			subject: __( 'New feedback received from your website', 'jetpack' ),
 		},
 	},
+
+	// This will take all child blocks
+	// and make then visible in the Block Library.
+	// The 'hiddenFromPicker' property will prevent
+	// them to show up in the Form placeholder
+	...map( childBlocks, block => ( {
+		name: block.name,
+		title: block.settings.title,
+		description: block.settings.description,
+		icon: block.settings.icon,
+		keywords: block.settings.keywords,
+		innerBlocks: [
+			[ `jetpack/${ block.name }`, { required: true } ],
+			[
+				'jetpack/button',
+				{
+					text: __( 'Submit', 'jetpack' ),
+					element: 'button',
+					lock: { remove: true },
+				},
+			],
+		],
+		hiddenFromPicker: true,
+	} ) ),
 ] );
 
 export default variations;