+++ projects/plugins/jetpack/tests/php/modules/publicize/test_class.publicize.php
@@ -1,5 +1,18 @@
 <?php
-require __DIR__ . '/../../../../modules/publicize.php';
+
+if ( ! function_exists( 'publicize_init' ) ) {
+	/**
+	 * Some tests rely on this function which won't get defined unless we mock lots
+	 * of things and require the module code. Instead we'll define it here.
+	 *
+	 * @return Publicize Object
+	 */
+	function publicize_init() {
+		global $publicize;
+
+		return $publicize;
+	}
+}
 
 /**
  * @group publicize