+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## 0.24.3 - 2022-11-01
+### Changed
+- Updated package dependencies.
+
+## 0.24.2 - 2022-10-25
+### Changed
+- Fix visual issues in the Product Price component in Jetpack plugin [#27032]
+
 ## 0.24.1 - 2022-10-19
 ### Changed
 - RNA: move product labels next to `legend` text [#26877]