+++ projects/js-packages/publicize-components/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.2] - 2022-11-01
+### Changed
+- Updated package dependencies. [#27196]
+
+## [0.8.1] - 2022-10-27
+### Fixed
+- Publicize Components: Fix the panel component refactor [#27095]
+
+## [0.8.0] - 2022-10-25
+### Added
+- Display broken connections to user in editor [#25803]
+
+### Changed
+- Reshare: Refactored the config logic and moved in the additional components for resharing [#25993]
+
+### Fixed
+- Social: Fix the path to the connections URL in the editor [#26932]
+
 ## [0.7.4] - 2022-10-17
 ### Changed
 - Updated package dependencies. [#26851]
@@ -109,6 +127,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#24470]
 
+[0.8.2]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.8.1...v0.8.2
+[0.8.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.8.0...v0.8.1
+[0.8.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.7.4...v0.8.0
 [0.7.4]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.7.3...v0.7.4
 [0.7.3]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.7.1...v0.7.2